third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/scroll-state/scroll-direction-mouse-drag-scroll-expected.txt

This is a testharness.js-based test.
[FAIL] Mouse drag scroll
  assert_equals: expected "yes" but got "no"
Harness: the test ran to completion.
